feat(repositories): change cardContainer name, loader
Change for a descriptive name of the component and loader custom in this section
