+++ frontend/src/Components/Repositories/Repositories.scss
@@ -1,3 +1,15 @@
+.repositories__loading {
+  align-items: center;
+  display:flex;
+  height: 50vh;
+  justify-content: center;
+
+  p {
+    color: var(--primary);
+    font-size: 18px;
+    font-weight: bold;
+  }
+}
 .card-container {
   border-bottom: 1px solid var(--border);
   display: flex;
